Reorganise code base and split `Server` into two.
The new `sable_network` crate contains definitions of network state
objects and a `Server` type that can link to the network, process and
propagate events, and emit a stream of `NetworkStateChange` objects for
application logic to process.

The IRC client protocol logic is moved to the new `sable_ircd` crate,
and the `ClientServer` type.