labeled-response: Avoid sending batches for single messages
This is not strictly necessary, but it is slightly annoying when manually
debugging from a client and wasted space on my terminal. :)

This replaces the `send_count` atomic counter with a `OnceLock` and
turns `sent_start` from a spinlock into a `Once` because we need to hold
the lock while the first message is being sent, now.