Make the network aware of user connections.
Users are no longer associated with a Server, because it's not a
one-to-one mapping and doesn't make sense to pretend it is. The server a
user is connected to - like the underlying real hostname - is a property
of the connection, is only shown to opers, and is shown for each active
connection.

This change also means that standard new-connection numerics are sent
when attaching to an existing session.